fix for 32-bit hash Closes #4433 Closes https://github.com/postgis/postgis/pull/429
git-svn-id: http://svn.osgeo.org/postgis/trunk@17575 b70326c6-7e19-0410-871a-916f4a2858ee

+++ liblwgeom/gserialized.c
@@ -110,7 +110,8 @@ uint32_t gserialized_max_header_size(void)
 * in the GSERIALIZED. Ignores metadata like flags and optional
 * boxes, etc.
 */
-uint64_t gserialized_hash(const GSERIALIZED *g)
+int32_t
+gserialized_hash(const GSERIALIZED *g)
 {
 	if (GFLAGS_GET_VERSION(g->gflags))
 		return gserialized2_hash(g);
